requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv  # Optional, for environment variable management
